Implement an obvious A*.
Kinda, because the algorithm as written needs to have cycles.

In particular, the way we represent the graph means that you need to go through
nodes without any progress in order to find the solution (the c<x> nodes).

The way to make it a proper A* (and more efficient) we need to flatten the graph
there so that the nodes without any storage don't appear on the graph...

Or something like that.